requirements_conda.txt
pandas
pytorch=1.3.1
torchvision=0.4.2
scipy
imageio
matplotlib
requests
tqdm
numpy=1.17.4
scikit-learn
networkx
numba=0.46.0
future
pyyaml
jinja2